+++ test/extensions/filters/http/router/config_test.cc
@@ -79,7 +79,7 @@ TEST(RouterFilterConfigTest, RouterV2Filter) {
   RouterFilterConfig factory;
   Http::FilterFactoryCb cb = factory.createFilterFactoryFromProto(router_config, "stats.", context);
   Http::MockFilterChainFactoryCallbacks filter_callback;
-  EXPECT_CALL(filter_callback, addStreamDecoderFilter(_));
+  EXPECT_CALL(filter_callback, addStreamDecoderFilter(_)).Times(1);
   cb(filter_callback);
 }
 
@@ -89,7 +89,7 @@ TEST(RouterFilterConfigTest, RouterFilterWithEmptyProtoConfig) {
   Http::FilterFactoryCb cb =
       factory.createFilterFactoryFromProto(*factory.createEmptyConfigProto(), "stats.", context);
   Http::MockFilterChainFactoryCallbacks filter_callback;
-  EXPECT_CALL(filter_callback, addStreamDecoderFilter(_));
+  EXPECT_CALL(filter_callback, addStreamDecoderFilter(_)).Times(1);
   cb(filter_callback);
 }